requirements.txt
fastapi
uvicorn[standard]
pydantic
pillow
pytesseract
requests
python-dotenv
sentence-transformers
vaderSentiment
tweepy
